requirements.txt
streamlit
langchain
langchain-core==1.0.0
langchain-text-splitters==1.0.0
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
pypdf
python-dotenv
pandas
openpyxl
